+++ src/layouts/Player/style.ts
@@ -3,4 +3,5 @@ import { SxProps } from "@mui/material"
 export const PaperSx: SxProps = {
 	p: "1rem",
 	mb: "1rem",
+	background:"#FBF9F9"
 }
